+++ test/blink/blink_test.exs
@@ -10,15 +10,15 @@ defmodule BlinkTest do
     end)
   end
 
-  describe "add_table/2" do
+  describe "with_table/2" do
     test "accepts atom and string table names" do
       defmodule Dummy do
         use Blink
 
         def call do
           new()
-          |> add_table(:atom)
-          |> add_table("string")
+          |> with_table(:atom)
+          |> with_table("string")
         end
 
         def table(_, _), do: []
@@ -34,8 +34,8 @@ defmodule BlinkTest do
 
         def call do
           new()
-          |> add_table("table_name")
-          |> add_table("table_name")
+          |> with_table("table_name")
+          |> with_table("table_name")
         end
 
         def table(_, _), do: []
@@ -47,15 +47,15 @@ defmodule BlinkTest do
     end
   end
 
-  describe "add_context/2" do
+  describe "with_context/2" do
     test "accepts atom and string keys" do
       defmodule Dummy do
         use Blink
 
         def call do
           new()
-          |> add_context(:atom)
-          |> add_context("string")
+          |> with_context(:atom)
+          |> with_context("string")
         end
 
         def context(_, _), do: []
@@ -71,8 +71,8 @@ defmodule BlinkTest do
 
         def call do
           new()
-          |> add_context("key")
-          |> add_context("key")
+          |> with_context("key")
+          |> with_context("key")
         end
 
         def context(_, _), do: []
@@ -84,23 +84,23 @@ defmodule BlinkTest do
     end
   end
 
-  describe "insert/2" do
+  describe "run/2" do
     test "can be overridden with custom implementation" do
       defmodule Dummy do
         use Blink
 
         def call do
           new()
-          |> add_table("users")
-          |> insert(BlinkTest.Repo)
+          |> with_table("users")
+          |> run(BlinkTest.Repo)
         end
 
         def table(_store, "users") do
           [%{id: 1, name: "Alice"}]
         end
 
-        def insert(store, repo) do
-          assert %Blink.Store{} = store
+        def run(seeder, repo) do
+          assert %Blink.Seeder{} = seeder
           assert BlinkTest.Repo = repo
 
           :some_custom_result
@@ -111,23 +111,23 @@ defmodule BlinkTest do
     end
   end
 
-  describe "insert/3" do
+  describe "run/3" do
     test "can be overridden with custom implementation" do
       defmodule Dummy do
         use Blink
 
         def call do
           new()
-          |> add_table("users")
-          |> insert(BlinkTest.Repo, batch_size: 500)
+          |> with_table("users")
+          |> run(BlinkTest.Repo, batch_size: 500)
         end
 
         def table(_store, "users") do
           [%{id: 1, name: "Alice"}]
         end
 
-        def insert(store, repo, opts) do
-          assert %Blink.Store{} = store
+        def run(seeder, repo, opts) do
+          assert %Blink.Seeder{} = seeder
           assert BlinkTest.Repo = repo
           assert [batch_size: 500] = opts
 
@@ -146,12 +146,12 @@ defmodule BlinkTest do
 
         def call do
           new()
-          |> add_table("users")
+          |> with_table("users")
         end
       end
 
       assert_raise ArgumentError,
-                   "you must define table/2 clauses that correspond with your calls to add_table/2",
+                   "you must define table/2 clauses that correspond with your calls to with_table/2",
                    fn ->
                      Dummy.call()
                    end
@@ -165,12 +165,12 @@ defmodule BlinkTest do
 
         def call do
           new()
-          |> add_context("data")
+          |> with_context("data")
         end
       end
 
       assert_raise ArgumentError,
-                   "you must define context/2 clauses that correspond with your calls to add_context/2",
+                   "you must define context/2 clauses that correspond with your calls to with_context/2",
                    fn ->
                      Dummy.call()
                    end